modals(fix): move to one folder
