+++ src/components/RecordTable.tsx
@@ -4,8 +4,8 @@ import styled from 'styled-components'
 import { FONT_SIZE } from '../constants'
 import { Record } from '../types'
 import BlockiesIdenticon from './Blockies'
-import SignatureModal from './SignatureModal'
-import TranscriptModal from './TranscriptModal'
+import SignatureModal from './modals/SignatureModal'
+import TranscriptModal from './modals/TranscriptModal'
 import { Trans, useTranslation } from 'react-i18next'
 import LoadingSpinner from './LoadingSpinner'
